refactor: move some inventory tests into subdirectory
